+++ src/resources/views/modules/create.blade.php
@@ -181,6 +181,10 @@
                     $selected_val = $('#category_id').val();
                     $('#media-wrapper').empty().load('medias/' + $selected_val);
                 }
+                loadMedia(); // Initiate media list on load
+                $(document).on('click', '.redmin-hierarchy li', function() {
+                    loadMedia(); // Load media list on category change
+                });
             })
         }(window.jQuery);
     </script>